+++ visualMidi/globalVariablesAndImports.pde
@@ -35,7 +35,7 @@ ArrayList<Note> sustainedNotes;
 float pitchBend = 0;
 float modulation = 0;
 float modulationRate = 0;
-int cutOffFilter = 0;
+float cutOffFilter = 0;
 float ampAtck;
 float ampDcy;
 float ampSus;
